Create separate webhook table with list helper
