+++ packages/api/src/store/db.ts
@@ -22,6 +22,7 @@ import QueueTable from "./queue";
 import { kebabToCamel } from "../util";
 import { QueryOptions, WithID } from "./types";
 import PushTargetTable from "./push-target-table";
+import WebhookTable from "./webhook-table";
 
 // Should be configurable, perhaps?
 const CONNECT_TIMEOUT = 5000;
@@ -45,7 +46,7 @@ export class DB {
   apiToken: Table<ApiToken>;
   user: Table<User>;
   usage: Table<Usage>;
-  webhook: Table<Webhook>;
+  webhook: WebhookTable;
   webhookResponse: Table<WebhookResponse>;
   passwordResetToken: Table<PasswordResetToken>;
   region: Table<Region>;
@@ -130,7 +131,7 @@ export class DB {
     });
     this.user = makeTable<User>({ db: this, schema: schemas["user"] });
     this.usage = makeTable<Usage>({ db: this, schema: schemas["usage"] });
-    this.webhook = makeTable<Webhook>({ db: this, schema: schemas["webhook"] });
+    this.webhook = new WebhookTable({ db: this, schema: schemas["webhook"] });
     this.passwordResetToken = makeTable<PasswordResetToken>({
       db: this,
       schema: schemas["password-reset-token"],
